Catch up with local wayland diff. use size_t for size, avoid casts.

+++ libswc/shm.c
@@ -47,7 +47,7 @@ static struct {
 struct pool {
 	struct wl_resource *resource;
 	void *data;
-	uint32_t size;
+	size_t size;
 	unsigned references;
 };
 
@@ -144,7 +144,7 @@ destroy(struct wl_client *client, struct wl_resource *resource)
 }
 
 static void
-resize(struct wl_client *client, struct wl_resource *resource, int32_t size)
+resize(struct wl_client *client, struct wl_resource *resource, size_t size)
 {
 	struct pool *pool = wl_resource_get_user_data(resource);
 	void *data;
@@ -167,7 +167,7 @@ static struct wl_shm_pool_interface shm_pool_implementation = {
 };
 
 static void
-create_pool(struct wl_client *client, struct wl_resource *resource, uint32_t id, int32_t fd, int32_t size)
+create_pool(struct wl_client *client, struct wl_resource *resource, uint32_t id, int32_t fd, size_t size)
 {
 	struct pool *pool;
